backend/requirements.txt
# python = 3.12.11

fastapi==0.124.0
uvicorn==0.24.0
pandas==2.1.3
gspread==5.12.0
google-auth==2.23.4
google-genai==1.52.0
python-dotenv==1.2.1
mysql-connector-python==9.5.0
sqlalchemy==2.0.45